--- a/store/types/gas_test.go
+++ b/store/types/gas_test.go
@@ -111,7 +111,7 @@ func TestAddUint64Overflow(t *testing.T) {
 func TestTransientGasConfig(t *testing.T) {
 	t.Parallel()
 	config := TransientGasConfig()
-	require.Equal(t, config, GasConfig{
+	require.Equal(t, config, &GasConfig{
 		HasCost:          100,
 		DeleteCost:       100,
 		ReadCostFlat:     100,