--- a/types/context_bench_test.go
+++ b/types/context_bench_test.go
@@ -1,0 +1,18 @@
+package types_test
+
+import (
+	"github.com/cosmos/cosmos-sdk/store/types"
+	"github.com/cosmos/cosmos-sdk/testutil"
+	"testing"
+)
+
+func BenchmarkContext_KVStore(b *testing.B) {
+	key := types.NewKVStoreKey(b.Name() + "_TestCacheContext")
+
+	ctx := testutil.DefaultContext(key, types.NewTransientStoreKey("transient_"+b.Name()))
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_ = ctx.KVStore(key)
+	}
+}