@testinginprod 2022-12-07
There's also something else, for instance applying a pointer receiver on `func (c *Context) KVStore` further cuts down execution times to `84ns` in my machine. I'm not sure if this change might cause some API breakages.