@aaronc 2022-12-07
Can we switch to pointer receivers here too @testinginprod ?