@peterbourgon 2022-12-08
Is the claim here that changing the gasConfig field of the KVStore struct from a value to a pointer represents a meaningful improvement in performance? If so, impressive! Can we see the benchstat deltas?