@testinginprod 2022-12-08
@peterbourgon yes sir... 
```
% benchstat 9bf027460bae724bbc1fddd3541248b7fe18ab2c.txt 8a9565fd0aca7909851fc90053e2cb8e25583761.txt
name                old time/op    new time/op    delta
Context_KVStore-10     192ns ± 1%      50ns ± 0%  -73.77%  (p=0.000 n=17+17)

name                old alloc/op   new alloc/op   delta
Context_KVStore-10     96.0B ± 0%     48.0B ± 0%  -50.00%  (p=0.000 n=20+20)

name                old allocs/op  new allocs/op  delta
Context_KVStore-10      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```

The two hashes refer to the commits former should be `main` latter should the last commit of this branch, respective benches:
main:
```
goos: darwin
goarch: arm64
pkg: github.com/cosmos/cosmos-sdk/types/bench
BenchmarkContext_KVStore-10    	 6275276	       189.7 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6307548	       190.6 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6283483	       191.7 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6259264	       191.3 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6257390	       191.5 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6279206	       191.7 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6243421	       191.4 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6250104	       191.5 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6258055	       193.0 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6301917	       192.6 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6252657	       192.8 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6187279	       192.1 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6263904	       192.6 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6172588	       245.2 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 5175621	       201.4 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6196754	       197.1 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6000902	       193.0 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6173347	       192.8 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6232633	       192.7 ns/op	      96 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	 6203295	       192.7 ns/op	      96 B/op	       1 allocs/op
PASS
ok  	github.com/cosmos/cosmos-sdk/types/bench	28.327s
```
current branch:
```
goos: darwin
goarch: arm64
pkg: github.com/cosmos/cosmos-sdk/types/bench
BenchmarkContext_KVStore-10    	22164160	        50.18 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23854467	        50.31 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23845243	        50.32 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23878081	        50.44 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23988165	        50.35 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23913512	        50.45 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23561230	        50.28 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23838060	        50.85 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23824905	        50.39 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23574750	        50.39 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23577394	        50.39 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23634318	        50.67 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23897776	        50.34 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23702649	        50.36 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	24019234	        50.37 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23486774	        50.54 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23950899	        50.39 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23440113	        50.29 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23687599	        51.03 ns/op	      48 B/op	       1 allocs/op
BenchmarkContext_KVStore-10    	23768612	        50.44 ns/op	      48 B/op	       1 allocs/op
PASS
ok  	github.com/cosmos/cosmos-sdk/types/bench	25.251s
```