@peterbourgon 2022-12-08
Neat! It seems like this benchmark is measuring the overhead of calling methods on value receivers vs. pointer receivers. It's definitely true that types of size larger than a pointer will copy more data during method invocations on value receivers vs. pointer receivers. Looks like in this case that's worth about 100ns per call to KVStore. But is that meaningful? What calls KVStore, and how often, and what does that call represent in the overall operation(s) that invoke it? Does subtracting 100ns from this specific call improve system performance in a measurable way? The PR modifies the semantics of the Context API, and introduces potential concurrency issues — are those modifications tested and ensured to be safe?