@testinginprod 2022-12-08
> are those modifications tested and ensured to be safe

This is a fair point, and I will add behavioural tests. But yeah the modifications made are safe.
The only cost is the fact that interfaces abstracting `sdk.Context` using `KVStore` and `TransientStore` methods will be broken (nothing was broken because of this API change in the SDK).

Then with regards to performance benefit, I think the change is trivial, easy to understand and easy to interpret the consequences (none, besides what mentioned above).

Now this is a starting point, I think; `sdk.Context` is passed everywhere as a value, which incurs in copy costs, the struct is too big to be passed around like this, and `sdk.Context` is core to every module, for instance:

- We use to emit events (have u checked how many events we create for one single TX? I did, and it's a LOOOT), every event emission is `ctx.EventManager()...`.
- We pass it in our chain of ante handlers, our ante handlers can pass it to 1 to 3-4 keepers (all ctx copies ofc, even if they shouldn't) then each of those keepers will call at least once `KVStore`.
- And TX execution hasn't even officially started.
- Then we start executing txs, we get again our context, pass it to keepers all around, call KVStore again... etc etc etc.
- Then we have post handlers...

Besides txs,
- there's begin block, everything involving keepers, and ctx being passed all around and functions calling methods on it...
- there's endblock, same as above.

I cannot provide you the exact amount of times ctx is passed around and the exact number of times methods are being called on it... But my biggest guess is: it's the most busy type in the SDK, I could even say more than store itself. (Not claiming context overhead > store overhead)